Fix QuickTime emulated OSErr type.
The OSErr type on Mac OS X is int16_t, not int32_t (see
http://developer.apple.com/mac/library/documentation/QuickTime/Reference/QTRef_DataTypes/Reference/reference.html).
The upper 16 bits will typically be something random (they're entirely
undefined). Change the type so it's right; a few places tried to compensate
for this by masking out the upper bits, but a few places also missed them,
which made for unpredictable behavior.


git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@30854 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ loader/qtx/qtxsdk/components.h
@@ -12,7 +12,7 @@ typedef unsigned char                   Boolean;
 typedef unsigned char                   Str31[32];
 typedef int32_t                         Fixed;
 
-typedef int32_t OSErr;
+typedef int16_t OSErr;
 typedef int OSType;
 
 typedef int32_t ComponentResult;
